Adjusted Mapper interface to provide World object access.
If running as a Minecraft server plugin, the Spigot/Bukkit API provides
direct access to much of the data that ChunkAtlas needs to generate
maps. MapCreator now checks if that access is available, and if so,
shares it with all Mapper objects by passing them a Bukkit World object
representing their mapped region.